Journal page added and updates migraines
