+++ public/assets/css/migraine_style.css
@@ -67,3 +67,7 @@ body {
   font-weight: bold;
   font-style: italic;
 }
+
+.nav-link {
+  font-size: 24px;
+}
