+++ views/partials/header.handlebars
@@ -1,8 +1,21 @@
 <nav class="navbar navbar-dark bg-dark">
-  <div class="col-md-7">
+  <div class="" align="left">
     <h1>Mígreni</h1>
   </div>
-  <div align="right" class="col-md-5">
-    <a href="/logout"><button class="btn btn-primary">Logout</button></a>
+  <div align="right" >
+  	<ul class="nav">
+	  	<li class="nav-item">
+		  	<a class="nav-link" href="/user">Home</a>
+		  </li>
+		  <li class="nav-item">
+		  	<a class="nav-link" href="/log">Journal</a>
+		  </li>
+		  <li class="nav-item">
+		  	<a class="nav-link" href="/chart">Chart</a>
+		  </li>
+		  	<li class="nav-item">
+		    <a class="nav-link" href="/logout"><button class="btn btn-primary">Logout</button></a>
+		  </li>
+	  </ul>
   </div>
 </nav>
