+++ views/survey.handlebars
@@ -1,33 +1,15 @@
-
+{{> header}}
 <div class="container-fluid">
-  <header class="row">
-    <h1>Mígreni</h1>
-  </header>
   <div class="row">
     <!-- sidebar -->
-    <div class="col-md-3 sidebar">
-      <br>
-      <!-- calendar -->
-      <div class="calendar"></div>
-      <br>
-      <!-- weather -->
-      <div class="weather card bg-light">
-        <div class="card-header">Current Weather</div>
-        <div class="card-body">
-          <h4 class="card-title" id="weather-city">Loading...</h4>
-          <p class="card-text temperature" id="weather-temp"></p>
-        </div>
-      </div>
-      <br>
-    </div>
-    <div class="col-md-5 content-main">
+    <div class="col-lg-9 content-main">
       <br><br>
       <div class="option-buttons">
         <button type="button" class="btn btn-dark btn-lg" id="show-survey">Log new migraine</button>
         <button type="button" class="btn btn-dark btn-lg" id="show-history">Show migraine history</button>
       </div>
       {{!-- history --}}
-      <div class="history">
+      <div class="history" hidden>
         <br>
         <h2>Your migraine history</h2>
 
@@ -58,7 +40,7 @@
         
       </div> {{!-- end history --}}
       <!-- survey -->
-      <div class="survey">
+      <div class="survey" hidden>
         <h2>Migraine questions</h2>
         <form>
           <br>
@@ -144,13 +126,20 @@
         </form>
       </div> {{!-- end survey --}}
     </div> {{!-- end content main --}}
-    <!-- charts aside -->
-    <div class="col-md-4 sidebar">
+    <div class="col-lg-3 sidebar">
       <br>
-      <canvas id="bar-chart-container" width="400" height="400"></canvas>
-      <img class="img-fluid" src="https://www.theatlas.com/i/atlas_H13grQLQ@2x.png" alt="Placeholder 1">
-      <br><br>
-      <img class="img-fluid" src="https://i.stack.imgur.com/qNiO7.png" alt="Placeholder 2">
+      <!-- weather -->
+      <div class="weather card bg-light">
+        <div class="card-header">Current Weather</div>
+        <div class="card-body">
+          <h4 class="card-title" id="weather-city">Loading...</h4>
+          <p class="card-text temperature" id="weather-temp"></p>
+        </div>
+      </div>
+      <br>
+      <!-- calendar -->
+      <div class="calendar"></div>
+      <br>
     </div>
   </div>
   {{!-- <footer class="row">
@@ -240,4 +229,5 @@
 
 <script src="/assets/js/pignose.calendar.js"></script>
 <script src="/assets/js/migraine.js"></script>
+<script src="/assets/js/updateMigraine.js"></script>
 {{!-- TODO MAKE THINGS UNIQUER --}}
